refactor: grade.repo→grade.repository로 변경 및 TxLike 공용 분리
