+++ src/common/prisma-tx.type.ts
@@ -1,0 +1,4 @@
+import { Prisma } from '@prisma/client';
+import { PrismaService } from 'src/prisma/prisma.service';
+
+export type TxLike = PrismaService | Prisma.TransactionClient;
